+++ lib/src/client/mcp_profile.dart
@@ -1,13 +1,43 @@
 import "client.dart";
+import "../../resources/fortnite_profile_ids.dart";
+import "../structures/profile_item.dart";
 
 /// mcp profile manager library
-class McpProfile {
+abstract class McpProfile {
   /// The master client
   late final Client client;
+
+  /// account id of the profile
+  late String accountId;
+
+  /// The profile id of mcp profile
+  late FortniteProfile profileId;
 
   /// is the profile initialized
   bool initialized = false;
 
+  /// When the profile was created
+  late DateTime created;
+
+  /// When the profile was last updated
+  late DateTime updated;
+
+  /// time of epic servers
+  late DateTime serverTime;
+
+  /// number of profile revisions
+  late int rvn;
+
+  /// profile items
+  List<ProfileItem> items = [];
+
   /// mcp profile object
-  McpProfile(this.client);
+  McpProfile(this.client, this.profileId) {
+    accountId = client.accountId;
+  }
+
+  bool confirmInitialized() {
+    if (initialized) return true;
+    throw Exception("Profile is not initialized yet");
+  }
 }